Module Openai

This module provides functions to interact with the OpenAI API for embeddings.

It includes functions to make HTTP POST requests to the OpenAI API embeddings endpoint and parse the response.

type embeddings_input = {

model : string;

input : string list;

}

Type definition for the input to the embeddings API.

val jsonaf_of_embeddings_input : embeddings_input -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val embeddings_input_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> embeddings_input

val sexp_of_embeddings_input : embeddings_input -> Sexplib0.Sexp.t

val embeddings_input_of_sexp : Sexplib0.Sexp.t -> embeddings_input

val bin_shape_embeddings_input : Bin_prot.Shape.t

val bin_size_embeddings_input : embeddings_input Bin_prot.Size.sizer

val bin_write_embeddings_input : embeddings_input Bin_prot.Write.writer

val bin_writer_embeddings_input : embeddings_input Bin_prot.Type_class.writer

val bin_read_embeddings_input : embeddings_input Bin_prot.Read.reader

val __bin_read_embeddings_input__ : (int -> embeddings_input) Bin_prot.Read.reader

val bin_reader_embeddings_input : embeddings_input Bin_prot.Type_class.reader

val bin_embeddings_input : embeddings_input Bin_prot.Type_class.t

type response = {

data : embedding list;

}

Type definition for the response from the embeddings API.

and embedding = {

embedding : float list;

index : int;

}

Type definition for an individual embedding in the response.

val jsonaf_of_response : response -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val jsonaf_of_embedding : embedding -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val response_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> response

val embedding_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> embedding

val sexp_of_response : response -> Sexplib0.Sexp.t

val sexp_of_embedding : embedding -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response

val embedding_of_sexp : Sexplib0.Sexp.t -> embedding

val bin_shape_response : Bin_prot.Shape.t

val bin_shape_embedding : Bin_prot.Shape.t

val bin_size_response : response Bin_prot.Size.sizer

val bin_write_response : response Bin_prot.Write.writer

val bin_writer_response : response Bin_prot.Type_class.writer

val bin_size_embedding : embedding Bin_prot.Size.sizer

val bin_write_embedding : embedding Bin_prot.Write.writer

val bin_writer_embedding : embedding Bin_prot.Type_class.writer

val bin_read_response : response Bin_prot.Read.reader

val __bin_read_response__ : (int -> response) Bin_prot.Read.reader

val bin_reader_response : response Bin_prot.Type_class.reader

val bin_read_embedding : embedding Bin_prot.Read.reader

val __bin_read_embedding__ : (int -> embedding) Bin_prot.Read.reader

val bin_reader_embedding : embedding Bin_prot.Type_class.reader

val bin_response : response Bin_prot.Type_class.t

val bin_embedding : embedding Bin_prot.Type_class.t

type func = {

name : string;

description : string option;

parameters : Jsonaf.t;

}

val jsonaf_of_func : func -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val func_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> func

val sexp_of_func : func -> Sexplib0.Sexp.t

val func_of_sexp : Sexplib0.Sexp.t -> func

val post_openai_embeddings : _ Eio.Net.t -> input:string list -> response

post_openai_embeddings ~input net makes an HTTP POST request to the OpenAI API embeddings endpoint with the given input and net.

It returns the parsed response as a response record.

type function_call = {

name : string;

arguments : string;

}

val jsonaf_of_function_call : function_call -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val function_call_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> function_call

val sexp_of_function_call : function_call -> Sexplib0.Sexp.t

val function_call_of_sexp : Sexplib0.Sexp.t -> function_call

val bin_shape_function_call : Bin_prot.Shape.t

val bin_size_function_call : function_call Bin_prot.Size.sizer

val bin_write_function_call : function_call Bin_prot.Write.writer

val bin_writer_function_call : function_call Bin_prot.Type_class.writer

val bin_read_function_call : function_call Bin_prot.Read.reader

val __bin_read_function_call__ : (int -> function_call) Bin_prot.Read.reader

val bin_reader_function_call : function_call Bin_prot.Type_class.reader

val bin_function_call : function_call Bin_prot.Type_class.t

type message = {

role : string;

content : string option;

name : string option;

function_call : function_call option;

}

val jsonaf_of_message : message -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val message_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> message

val sexp_of_message : message -> Sexplib0.Sexp.t

val message_of_sexp : Sexplib0.Sexp.t -> message

val bin_shape_message : Bin_prot.Shape.t

val bin_size_message : message Bin_prot.Size.sizer

val bin_write_message : message Bin_prot.Write.writer

val bin_writer_message : message Bin_prot.Type_class.writer

val bin_read_message : message Bin_prot.Read.reader

val __bin_read_message__ : (int -> message) Bin_prot.Read.reader

val bin_reader_message : message Bin_prot.Type_class.reader

val bin_message : message Bin_prot.Type_class.t

type delta = {

role : string option;

content : string option;

function_call : function_call option;

}

val jsonaf_of_delta : delta -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val delta_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> delta

val sexp_of_delta : delta -> Sexplib0.Sexp.t

val delta_of_sexp : Sexplib0.Sexp.t -> delta

val bin_shape_delta : Bin_prot.Shape.t

val bin_size_delta : delta Bin_prot.Size.sizer

val bin_write_delta : delta Bin_prot.Write.writer

val bin_writer_delta : delta Bin_prot.Type_class.writer

val bin_read_delta : delta Bin_prot.Read.reader

val __bin_read_delta__ : (int -> delta) Bin_prot.Read.reader

val bin_reader_delta : delta Bin_prot.Type_class.reader

val bin_delta : delta Bin_prot.Type_class.t

type choice = {

delta : delta;

finish_reason : string option;

}

val jsonaf_of_choice : choice -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val choice_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> choice

val sexp_of_choice : choice -> Sexplib0.Sexp.t

val choice_of_sexp : Sexplib0.Sexp.t -> choice

val bin_shape_choice : Bin_prot.Shape.t

val bin_size_choice : choice Bin_prot.Size.sizer

val bin_write_choice : choice Bin_prot.Write.writer

val bin_writer_choice : choice Bin_prot.Type_class.writer

val bin_read_choice : choice Bin_prot.Read.reader

val __bin_read_choice__ : (int -> choice) Bin_prot.Read.reader

val bin_reader_choice : choice Bin_prot.Type_class.reader

val bin_choice : choice Bin_prot.Type_class.t

type message_response = {

message : delta;

finish_reason : string option;

}

val jsonaf_of_message_response : message_response -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val message_response_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> message_response

val sexp_of_message_response : message_response -> Sexplib0.Sexp.t

val message_response_of_sexp : Sexplib0.Sexp.t -> message_response

val bin_shape_message_response : Bin_prot.Shape.t

val bin_size_message_response : message_response Bin_prot.Size.sizer

val bin_write_message_response : message_response Bin_prot.Write.writer

val bin_writer_message_response : message_response Bin_prot.Type_class.writer

val bin_read_message_response : message_response Bin_prot.Read.reader

val __bin_read_message_response__ : (int -> message_response) Bin_prot.Read.reader

val bin_reader_message_response : message_response Bin_prot.Type_class.reader

val bin_message_response : message_response Bin_prot.Type_class.t

type _ response_type =

| Stream : (choice -> unit) -> unit response_type

| Default : message_response response_type

type model =

| Gpt4

| Gpt3

| Gpt3_16k

post_chat_completion ?max_tokens ?stream env inputs makes an HTTP POST request to the OpenAI API chat completion endpoint with the given optional max_tokens, optional stream, env, and inputs.

It returns the parsed response as a string.

val post_chat_completion : 'a. 'a response_type -> ?max_tokens:int -> ?temperature:float -> ?functions:func list -> ?model:model -> 'n Eio.Net.t -> inputs:message list -> 'a