Module Template.Make_Template

The Make functor creates a templating module for a given RENDERABLE type. Supports variable replacement with the syntax {variable

}

Parameters

module R : RENDERABLE

Signature

type t

val create : string -> t

create s creates a new template with the given string s.

val render : t -> R.t -> string

render t r renders the template t with the RENDERABLE data r.

val to_string : t -> string

to_string t converts the template t to a string.